move dict_add from config_sampling to utils
