+++ utils.py
@@ -1,4 +1,6 @@
+import copy
 import tensorflow as tf
+
 
 class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
     def __init__(self, d_model, decay):
@@ -17,6 +19,18 @@ class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
 def safe_div(x, y, eps=1e-8):
     # returns safe x / max(y, epsilon)
     return x / tf.maximum(y, eps)
+
+
+def dict_add(first: dict, second: dict):
+    output = copy.deepcopy(first)
+
+    for key in second.keys():
+        if key in output:
+            output[key] += second[key]
+        else:
+            output[key] = second[key]
+
+    return output
 
 
 def get_device():
